[common,PAL/Linux-SGX] Move IOCTLs from Linux-SGX to common code
This internal change in Gramine makes flexible IOCTLs generic for all
PALs. A future commit will use these new IOCTL APIs in Linux PAL.
