+++ common/src/ioctls/meson.build
@@ -1,0 +1,12 @@
+ioctls_inc = [
+    include_directories('.'),
+    common_inc,
+]
+
+ioctls_dep = declare_dependency(
+    sources: [
+        'ioctls.c',
+        'ioctls.h',
+    ],
+    include_directories: ioctls_inc,
+)
